Move hpx/util/zip_iterator.hpp to segmented_algorithms module
Specializes traits for segmented algorithms.
