+++ libs/include/include/hpx/include/util.hpp
@@ -21,6 +21,7 @@
 #include <hpx/iterator_support/iterator_adaptor.hpp>
 #include <hpx/iterator_support/iterator_facade.hpp>
 #include <hpx/pack_traversal/unwrap.hpp>
+#include <hpx/parallel/segmented_algorithms/traits/zip_iterator.hpp>
 #include <hpx/preprocessor/cat.hpp>
 #include <hpx/preprocessor/expand.hpp>
 #include <hpx/preprocessor/nargs.hpp>
@@ -37,6 +38,5 @@
 #include <hpx/util/interval_timer.hpp>
 #include <hpx/util/thread_aware_timer.hpp>
 #include <hpx/util/to_string.hpp>
-#include <hpx/util/zip_iterator.hpp>
 
 #endif
