+++ libs/resource_partitioner/examples/async_customization.cpp
@@ -4,6 +4,7 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#include <hpx/hpx_init.hpp>
 #include <hpx/include/parallel_executors.hpp>
 #include <hpx/thread_executors/default_executor.hpp>
 #include <hpx/thread_executors/pool_executor.hpp>
