+++ libs/segmented_algorithms/tests/unit/partitioned_vector_transform_reduce1.cpp
@@ -8,9 +8,9 @@
 #include <hpx/include/parallel_for_each.hpp>
 #include <hpx/include/parallel_transform_reduce.hpp>
 #include <hpx/include/partitioned_vector_predef.hpp>
+#include <hpx/parallel/segmented_algorithms/traits/zip_iterator.hpp>
 #include <hpx/parallel/util/zip_iterator.hpp>
 #include <hpx/testing.hpp>
-#include <hpx/util/zip_iterator.hpp>
 
 #include <iterator>
